The DangerousGetHandle() in getRegistryHandle() is no longer casted to IntPtr (Thanks to VSeven)

+++ Little Registry Cleaner/XML Registry/xmlRegistry.cs
@@ -106,7 +106,8 @@ namespace Little_Registry_Cleaner.Xml
             Type type = registryKey.GetType();
             FieldInfo fieldInfo = type.GetField("hkey", BindingFlags.Instance | BindingFlags.NonPublic);
             System.Runtime.InteropServices.SafeHandle i = (System.Runtime.InteropServices.SafeHandle)fieldInfo.GetValue(registryKey);
-            return ((IntPtr)i.DangerousGetHandle()).ToInt32();
+
+            return i.DangerousGetHandle().ToInt32();
         }
 
         ArrayList addRegistryValues(int hKey)
